+++ src/page/search/competition/index.tsx
@@ -1,7 +1,15 @@
-import Search from "feature/search/index"
+import SearchFormContainer from "feature/search/container/searchFormContainer"
+import SearchListContainer from "feature/search/container/searchListContainer"
+import PagenationContainer from "feature/search/container/pagenationContainer"
 
 function SearchCompetition(): JSX.Element {
-  return <Search></Search>
+  return (
+    <>
+      <SearchFormContainer />
+      <SearchListContainer />
+      <PagenationContainer />
+    </>
+  )
 }
 
 export default SearchCompetition
